refactor(mpp): rename walker entry to distribute_plan; retire exec_bridge and shape builders
Aligns the MPP walker's public name with the DF-D reference (`distribute_plan`
takes a standard physical plan and produces a distributed one), and retires
the two modules that were only alive to forward into the walker:

* `exec_bridge.rs` was a 6-line alias over `walker::annotate_plan`; its
  three `build_*_bridge` functions and supporting helpers duplicated the
  walker's topology builders and were never called.
* `shape.rs`'s `build_scalar_agg_on_binary_join` / `build_groupby_agg_on_binary_join`
  / `build_join_only` and their `*Inputs` structs were consumed only by the
  dead bridges above; kept the classifier (`MppPlanShape`, `ClassifyInputs`,
  `classify`) and added six unit tests covering its truth table.

Callers (`aggregatescan::exec_custom_scan`, `joinscan::exec_custom_scan`)
now call `walker::distribute_plan` directly. Regression expected-outputs
for `mpp_fallback` and `mpp_exec` updated to match the renamed log prefix.

Net -566 LOC; no functional change. Sets up a verbatim DF-D-structured
generic cut walker as the next step.
