+++ pg_search/src/postgres/customscan/mpp/shape.rs
@@ -15,23 +15,22 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-//! MPP plan-shape classifier + per-shape physical-plan builders.
+//! MPP plan-shape classifier.
 //!
-//! Different query shapes want different MPP topologies — trying to force one
-//! plan structure across all of them breaks correctness (the earlier
-//! `build_mpp_aggregate_plan` baked in a post-Partial-shuffle topology that
-//! produced spurious rows for scalar aggregation). This module classifies the
-//! query, dispatches to the right builder, and each builder composes only the
-//! nodes it needs.
+//! Classifies a query (at planning time) into one of a handful of MPP
+//! topology shapes so `aggregatescan` / `joinscan` can decide whether the
+//! plan is MPP-eligible and, if so, pick the right DSM mesh allocation size.
+//! The actual plan-rewriting lives in [`super::walker::distribute_plan`],
+//! which derives the same shape from plan structure and uses the
+//! classifier's output only as a sanity cross-check.
 //!
 //! # Supported shapes
 //!
 //! [`MppPlanShape::ScalarAggOnBinaryJoin`] — e.g., `SELECT COUNT(*) FROM f
 //! JOIN p WHERE ...`. Per worker: pre-join shuffle of each side → HashJoin →
-//! AggregateExec(Partial). Partial rows are shipped to PG's Gather and
-//! finalized by PG's native `Finalize Aggregate` above. No post-Partial
-//! shuffle is inserted (it would produce spurious all-NULL rows for scalar
-//! aggregation on non-hit workers).
+//! AggregateExec(Partial). Partial rows are shipped to seat 0 and finalized
+//! by a single `AggregateExec(FinalPartitioned)` on the leader; workers
+//! emit zero rows so PG's Gather sees exactly one row per query.
 //!
 //! [`MppPlanShape::GroupByAggOnBinaryJoin`] — e.g., `SELECT k, COUNT(*) FROM
 //! f JOIN p GROUP BY k`. Per worker: pre-join shuffle → HashJoin →
@@ -42,46 +41,13 @@
 //!
 //! [`MppPlanShape::GroupByAggSingleTable`] — `SELECT k, COUNT(*) FROM t
 //! GROUP BY k`. No join; Partial → post-Partial shuffle → FinalPartitioned.
+//! Not yet plumbed through the walker; classifier accepts but the walker
+//! returns `DataFusionError::Plan`.
 //!
 //! [`MppPlanShape::JoinOnly`] — `SELECT ... FROM f JOIN p`. Pre-join shuffle
 //! → HashJoin → emit rows (no aggregate).
 //!
 //! [`MppPlanShape::Ineligible`] — fall back to the non-MPP serial path.
-//!
-//! # What this module *does not* do
-//!
-//! The DSM hook layer allocates the mesh wirings. A single shuffle =
-//! one mesh. Binary join under `ScalarAggOnBinaryJoin` or
-//! `GroupByAggOnBinaryJoin` needs TWO meshes (one per join input).
-//! `GroupByAggOnBinaryJoin` adds a THIRD mesh for the post-Partial
-//! shuffle. Multi-mesh DSM allocation is a separate piece — this module
-//! takes pre-allocated wirings as parameters.
-
-// caller lands once create_custom_path flips parallel-safe
-// `CoalesceBatchesExec` is deprecated in favor of arrow-rs's `BatchCoalescer`,
-// but DataFusion 52 still ships it as an `ExecutionPlan` node — the replacement
-// is a streaming coalescer, not a plan node, so we cannot drop into the same
-// slot. Remove this allow when the wrapper migrates.
-#![allow(deprecated)]
-
-use std::sync::Arc;
-
-use datafusion::arrow::datatypes::SchemaRef;
-use datafusion::common::Result;
-use datafusion::logical_expr::JoinType;
-use datafusion::physical_expr::aggregate::AggregateFunctionExpr;
-use datafusion::physical_expr::PhysicalExpr;
-use datafusion::physical_plan::aggregates::{AggregateExec, AggregateMode, PhysicalGroupBy};
-use datafusion::physical_plan::coalesce_batches::CoalesceBatchesExec;
-use datafusion::physical_plan::filter::FilterExec;
-use datafusion::physical_plan::joins::utils::JoinOn;
-use datafusion::physical_plan::joins::{HashJoinExec, PartitionMode};
-use datafusion::physical_plan::ExecutionPlan;
-
-use crate::postgres::customscan::mpp::plan_build::{wrap_with_mpp_shuffle, MppShuffleInputs};
-use crate::postgres::customscan::mpp::shuffle::ShuffleWiring;
-use crate::postgres::customscan::mpp::stage::MppStage;
-use crate::postgres::customscan::mpp::transport::DrainHandle;
 
 /// Classify a query so the dispatcher picks the right topology.
 #[derive(Debug, Clone, Copy, Eq, PartialEq)]
@@ -139,405 +105,29 @@ pub fn classify(inputs: &ClassifyInputs) -> MppPlanShape {
     }
 }
 
-// ============================================================================
-// Per-shape plan builders
-// ============================================================================
-
-/// Inputs to [`build_scalar_agg_on_binary_join`]. Both `left_child` and
-/// `right_child` are the pre-shuffle scans (scan → filter) for each join
-/// input. The wirings + drains are the caller-allocated halves of three
-/// independent meshes: one per join input plus a final-gather mesh that
-/// routes every worker's Partial row to the leader seat so a single
-/// `AggregateExec(FinalPartitioned)` on the leader produces the one-row
-/// scalar result.
-pub struct ScalarAggOnBinaryJoinInputs {
-    pub left_child: Arc<dyn ExecutionPlan>,
-    pub right_child: Arc<dyn ExecutionPlan>,
-    pub left_shuffle: ShuffleWiring,
-    pub left_drain: std::sync::Arc<DrainHandle>,
-    pub left_schema: SchemaRef,
-    pub right_shuffle: ShuffleWiring,
-    pub right_drain: std::sync::Arc<DrainHandle>,
-    pub right_schema: SchemaRef,
-    pub aggr_expr: Vec<Arc<AggregateFunctionExpr>>,
-    pub filter_expr: Vec<Option<Arc<dyn PhysicalExpr>>>,
-    /// Wiring for the final-gather mesh: every participant ships its
-    /// Partial row to one fixed target seat (the leader, seat 0) via a
-    /// [`FixedTargetPartitioner`](super::shuffle::FixedTargetPartitioner).
-    /// The caller must build this with `target == 0` so PG's Gather sees
-    /// exactly one row (from the leader) per query.
-    pub final_shuffle: ShuffleWiring,
-    pub final_drain: std::sync::Arc<DrainHandle>,
-    /// The original `HashJoinExec` from the standard plan. Rebuilt here via
-    /// `builder().with_new_children(...)` so its `dynamic_filter` Arc (populated
-    /// by `SharedBuildAccumulator` once the local build side completes) is
-    /// preserved across the MPP rewrite. The bridge strips the Arc from the
-    /// probe-side `PgSearchScanPlan` and this module re-applies it as a
-    /// [`FilterExec`] on top of the post-shuffle probe stream (see below).
-    pub original_hash_join: Arc<HashJoinExec>,
-    /// Stage descriptor for the left pre-join mesh (stage 0 by convention).
-    pub left_stage: MppStage,
-    /// Stage descriptor for the right pre-join mesh (stage 1 by convention).
-    pub right_stage: MppStage,
-    /// Stage descriptor for the gather-to-leader final mesh (stage 2 by
-    /// convention). Every participant ships its Partial row on this stage.
-    pub final_stage: MppStage,
-}
-
-/// Build the MPP plan for a scalar aggregate over a binary join.
-///
-/// # Leader / worker asymmetry
-///
-/// For a scalar aggregate, DataFusion's `AggregateExec(FinalPartitioned)`
-/// emits *one* row even when its input is empty (the SQL semantics for
-/// `SELECT COUNT(*) FROM empty` is `0`, not "no row"). If every
-/// participant ran the full plan, PG's Gather above would concatenate N
-/// rows (one per worker + leader), breaking the scalar contract.
-///
-/// Instead, the leader and workers build asymmetric plans driven by
-/// `is_leader`:
-///
-/// **Leader** (is_leader = true):
-/// ```text
-///     AggregateExec(FinalPartitioned)
-///       wrap_with_mpp_shuffle(final_mesh, all-to-seat-0)
-///         AggregateExec(Partial)
-///           HashJoinExec(PartitionMode::Partitioned)
-///             wrap_with_mpp_shuffle(left_child,  left_wiring, left_drain)
-///             wrap_with_mpp_shuffle(right_child, right_wiring, right_drain)
-/// ```
-/// Leader's own Partial row stays local (self-partition == target 0);
-/// its drain receives Partial rows from every worker. `FinalPartitioned`
-/// sums N partials → emits one row.
-///
-/// **Worker** (is_leader = false):
-/// ```text
-///     wrap_with_mpp_shuffle(final_mesh, all-to-seat-0)   ← no FinalPartitioned
-///       AggregateExec(Partial)
-///         HashJoinExec(PartitionMode::Partitioned)
-///           wrap_with_mpp_shuffle(left_child,  …)
-///           wrap_with_mpp_shuffle(right_child, …)
-/// ```
-/// Worker's ShuffleExec ships its Partial row to seat 0 via the final
-/// mesh (self != target so its self-partition is empty). Worker's
-/// DrainGatherExec reads from the worker's inbound receivers on the
-/// final mesh, which are empty by construction (everyone ships *to*
-/// seat 0, nobody ships *to* the worker). The worker's stream therefore
-/// emits zero rows — PG's Gather sees exactly one row per query (from
-/// the leader).
-pub fn build_scalar_agg_on_binary_join(
-    inputs: ScalarAggOnBinaryJoinInputs,
-    is_leader: bool,
-) -> Result<Arc<dyn ExecutionPlan>> {
-    let left_shuffled = wrap_with_mpp_shuffle(MppShuffleInputs {
-        child: inputs.left_child,
-        wiring: inputs.left_shuffle,
-        drain_handle: inputs.left_drain,
-        wrapped_schema: inputs.left_schema,
-        tag: "scalar_left",
-        stage: Some(inputs.left_stage),
-    })?;
-    let right_shuffled = wrap_with_mpp_shuffle(MppShuffleInputs {
-        child: inputs.right_child,
-        wiring: inputs.right_shuffle,
-        drain_handle: inputs.right_drain,
-        wrapped_schema: inputs.right_schema,
-        tag: "scalar_right",
-        stage: Some(inputs.right_stage),
-    })?;
-
-    // Re-apply the HashJoin's dynamic filter above the post-shuffle probe
-    // stream. The filter was stripped from the probe-side scan by the bridge
-    // so it isn't applied before shuffle routing (which would drop rows
-    // destined for peer participants); re-attaching it here lets each
-    // participant's local `SharedBuildAccumulator` populate the Arc once its
-    // local build side finishes, and the `FilterExec` then prunes probe rows
-    // that cannot match any local build key.
-    let right_probe: Arc<dyn ExecutionPlan> =
-        match inputs.original_hash_join.dynamic_filter_for_test().cloned() {
-            Some(df) => Arc::new(FilterExec::try_new(df, right_shuffled)?),
-            None => right_shuffled,
-        };
-
-    // Rebuild through the builder so the `dynamic_filter` field survives. A
-    // plain `HashJoinExec::try_new` would clear it (the builder initializes
-    // `dynamic_filter: None` and `try_new` never calls `with_dynamic_filter`),
-    // orphaning the Arc held by the `FilterExec` above.
-    let join = inputs
-        .original_hash_join
-        .builder()
-        .with_new_children(vec![left_shuffled, right_probe])?
-        .build_exec()?;
-    let join_schema = join.schema();
-
-    let partial: Arc<dyn ExecutionPlan> = Arc::new(AggregateExec::try_new(
-        AggregateMode::Partial,
-        PhysicalGroupBy::new_single(vec![]),
-        inputs.aggr_expr.clone(),
-        inputs.filter_expr.clone(),
-        join,
-        join_schema,
-    )?);
-    let partial_schema = partial.schema();
-
-    let gathered = wrap_with_mpp_shuffle(MppShuffleInputs {
-        child: partial,
-        wiring: inputs.final_shuffle,
-        drain_handle: inputs.final_drain,
-        wrapped_schema: partial_schema.clone(),
-        tag: "scalar_final",
-        stage: Some(inputs.final_stage),
-    })?;
-
-    if !is_leader {
-        // Worker plan: stream drives the shuffle (shipping Partial to
-        // the leader) and returns zero rows locally.
-        return Ok(gathered);
-    }
-
-    let final_agg = AggregateExec::try_new(
-        AggregateMode::FinalPartitioned,
-        PhysicalGroupBy::new_single(vec![]),
-        inputs.aggr_expr,
-        inputs.filter_expr,
-        gathered,
-        partial_schema,
-    )?;
-    Ok(Arc::new(final_agg))
-}
-
-/// Inputs to [`build_groupby_agg_on_binary_join`]. Adds a third mesh
-/// (`postagg_shuffle` + `postagg_drain`) on top of
-/// [`ScalarAggOnBinaryJoinInputs`] for the Partial → FinalPartitioned
-/// shuffle on group keys. Caller supplies group-by expressions +
-/// per-partial schema.
-pub struct GroupByAggOnBinaryJoinInputs {
-    pub left_child: Arc<dyn ExecutionPlan>,
-    pub right_child: Arc<dyn ExecutionPlan>,
-    pub left_shuffle: ShuffleWiring,
-    pub left_drain: std::sync::Arc<DrainHandle>,
-    pub left_schema: SchemaRef,
-    pub right_shuffle: ShuffleWiring,
-    pub right_drain: std::sync::Arc<DrainHandle>,
-    pub right_schema: SchemaRef,
-    pub group_by: PhysicalGroupBy,
-    pub aggr_expr: Vec<Arc<AggregateFunctionExpr>>,
-    pub filter_expr: Vec<Option<Arc<dyn PhysicalExpr>>>,
-    pub postagg_shuffle: ShuffleWiring,
-    pub postagg_drain: std::sync::Arc<DrainHandle>,
-    /// See [`ScalarAggOnBinaryJoinInputs::original_hash_join`].
-    pub original_hash_join: Arc<HashJoinExec>,
-    /// Stage descriptor for the left pre-join mesh (stage 0 by convention).
-    pub left_stage: MppStage,
-    /// Stage descriptor for the right pre-join mesh (stage 1 by convention).
-    pub right_stage: MppStage,
-    /// Stage descriptor for the Partial→FinalPartitioned post-aggregate
-    /// shuffle mesh (stage 2 by convention).
-    pub postagg_stage: MppStage,
-}
-
-/// Build the MPP plan for a GROUP BY aggregate over a binary join.
-///
-/// Symmetric across all seats — no leader/worker asymmetry. Each seat
-/// runs the same plan; rows land on the seat that owns their group-by
-/// hash partition, so `FinalPartitioned` on each seat emits a disjoint
-/// set of groups and PG's Gather concatenates.
-///
-/// Shape (every seat):
-/// ```text
-///     AggregateExec(FinalPartitioned, group_by)
-///       wrap_with_mpp_shuffle(gb_postagg, HashPartitioner(group_keys))
-///         CoalesceBatchesExec(target = 64 Ki rows)
-///           AggregateExec(Partial, group_by)
-///             HashJoinExec(Partitioned)
-///               wrap_with_mpp_shuffle(gb_left)
-///               wrap_with_mpp_shuffle(gb_right)
-/// ```
-///
-/// # Why `CoalesceBatchesExec` before the post-agg shuffle
-///
-/// `AggregateExec(Partial)` emits DataFusion's default 8 Ki-row batches.
-/// At 25 M input / 3.12 M distinct groups, that's ~191 batches per seat
-/// going over shm_mq via Arrow-IPC. Per-batch overhead (schema/dictionary
-/// preamble + per-iteration dispatch in `ShuffleStream::process_batch`)
-/// dominates at small batch sizes. Coalescing to 64 Ki rows collapses
-/// that to ~24 batches, amortizing the fixed per-batch cost ~8× and
-/// keeping payload under the 64 MiB `shm_mq` queue capacity so
-/// backpressure stays near zero.
-///
-/// # Why hash-partition (not gather-to-leader) once coalesce lands
-///
-/// Hash-partitioning by group keys lets each seat finalize 1/N of the
-/// distinct groups in parallel; the gather-to-leader variant
-/// (earlier `FixedTargetPartitioner(0)`) serialized `Final` on the
-/// leader. Gather was a win only while per-batch encode cost was the
-/// binding constraint — with batches coalesced, encode is cheap enough
-/// that parallel `Final` dominates.
-pub fn build_groupby_agg_on_binary_join(
-    inputs: GroupByAggOnBinaryJoinInputs,
-) -> Result<Arc<dyn ExecutionPlan>> {
-    let left_shuffled = wrap_with_mpp_shuffle(MppShuffleInputs {
-        child: inputs.left_child,
-        wiring: inputs.left_shuffle,
-        drain_handle: inputs.left_drain,
-        wrapped_schema: inputs.left_schema,
-        tag: "gb_left",
-        stage: Some(inputs.left_stage),
-    })?;
-    let right_shuffled = wrap_with_mpp_shuffle(MppShuffleInputs {
-        child: inputs.right_child,
-        wiring: inputs.right_shuffle,
-        drain_handle: inputs.right_drain,
-        wrapped_schema: inputs.right_schema,
-        tag: "gb_right",
-        stage: Some(inputs.right_stage),
-    })?;
-
-    // See `build_scalar_agg_on_binary_join` for the rationale behind this
-    // post-shuffle `FilterExec` + `builder()` rebuild.
-    let right_probe: Arc<dyn ExecutionPlan> =
-        match inputs.original_hash_join.dynamic_filter_for_test().cloned() {
-            Some(df) => Arc::new(FilterExec::try_new(df, right_shuffled)?),
-            None => right_shuffled,
-        };
-
-    let join = inputs
-        .original_hash_join
-        .builder()
-        .with_new_children(vec![left_shuffled, right_probe])?
-        .build_exec()?;
-    let join_schema = join.schema();
-
-    let partial: Arc<dyn ExecutionPlan> = Arc::new(AggregateExec::try_new(
-        AggregateMode::Partial,
-        inputs.group_by.clone(),
-        inputs.aggr_expr.clone(),
-        inputs.filter_expr.clone(),
-        join,
-        join_schema,
-    )?);
-    // Coalesce Partial's default 8 Ki-row batches into 64 Ki-row batches
-    // before the post-agg shuffle so Arrow-IPC encode amortizes over
-    // ~8× fewer batches. See the function-level doc for why.
-    let coalesced_partial: Arc<dyn ExecutionPlan> =
-        Arc::new(CoalesceBatchesExec::new(partial, 65_536));
-    let partial_schema = coalesced_partial.schema();
-
-    let repartitioned = wrap_with_mpp_shuffle(MppShuffleInputs {
-        child: coalesced_partial,
-        wiring: inputs.postagg_shuffle,
-        drain_handle: inputs.postagg_drain,
-        wrapped_schema: partial_schema.clone(),
-        tag: "gb_postagg",
-        stage: Some(inputs.postagg_stage),
-    })?;
-
-    let final_agg = AggregateExec::try_new(
-        AggregateMode::FinalPartitioned,
-        inputs.group_by,
-        inputs.aggr_expr,
-        inputs.filter_expr,
-        repartitioned,
-        partial_schema,
-    )?;
-    Ok(Arc::new(final_agg))
-}
-
-/// Inputs to [`build_join_only`]. Two meshes (one per join input).
-pub struct JoinOnlyInputs {
-    pub left_child: Arc<dyn ExecutionPlan>,
-    pub right_child: Arc<dyn ExecutionPlan>,
-    pub left_shuffle: ShuffleWiring,
-    pub left_drain: std::sync::Arc<DrainHandle>,
-    pub left_schema: SchemaRef,
-    pub right_shuffle: ShuffleWiring,
-    pub right_drain: std::sync::Arc<DrainHandle>,
-    pub right_schema: SchemaRef,
-    pub join_on: JoinOn,
-    /// Column projection applied to the `HashJoinExec` output. DataFusion
-    /// may prune the raw `left_schema ++ right_schema` down to only the
-    /// columns the caller needs; forwarding the same projection here keeps
-    /// downstream column indices valid.
-    pub join_projection: Option<Vec<usize>>,
-    pub join_type: JoinType,
-    /// Stage descriptor for the left pre-join mesh (stage 0 by convention).
-    pub left_stage: MppStage,
-    /// Stage descriptor for the right pre-join mesh (stage 1 by convention).
-    pub right_stage: MppStage,
-}
-
-/// Build the MPP plan for a join without an aggregate on top. Emits the
-/// join output rows directly.
-///
-/// Shape:
-/// ```text
-///     HashJoinExec(PartitionMode::Partitioned)
-///       wrap_with_mpp_shuffle(left_child,  hash=[left_key])
-///       wrap_with_mpp_shuffle(right_child, hash=[right_key])
-/// ```
-pub fn build_join_only(inputs: JoinOnlyInputs) -> Result<Arc<dyn ExecutionPlan>> {
-    let left_shuffled = wrap_with_mpp_shuffle(MppShuffleInputs {
-        child: inputs.left_child,
-        wiring: inputs.left_shuffle,
-        drain_handle: inputs.left_drain,
-        wrapped_schema: inputs.left_schema,
-        tag: "join_left",
-        stage: Some(inputs.left_stage),
-    })?;
-    let right_shuffled = wrap_with_mpp_shuffle(MppShuffleInputs {
-        child: inputs.right_child,
-        wiring: inputs.right_shuffle,
-        drain_handle: inputs.right_drain,
-        wrapped_schema: inputs.right_schema,
-        tag: "join_right",
-        stage: Some(inputs.right_stage),
-    })?;
-
-    let join = HashJoinExec::try_new(
-        left_shuffled,
-        right_shuffled,
-        inputs.join_on,
-        None,
-        &inputs.join_type,
-        inputs.join_projection,
-        PartitionMode::Partitioned,
-        datafusion::common::NullEquality::NullEqualsNothing,
-        false,
-    )?;
-    Ok(Arc::new(join))
-}
-
-// ============================================================================
-// Cut counting moved to `walker::cut_count_for_shape` (P4). The walker is the
-// single source of truth for how many shuffles a shape produces, since the
-// DSM-estimate hooks now size the region from that helper.
-// ============================================================================
-
 #[cfg(test)]
 mod tests {
     use super::*;
 
-    fn inputs(n: usize, group: bool, agg: bool, splittable: bool) -> ClassifyInputs {
+    fn inputs(n: usize, agg: bool, gb: bool, splittable: bool) -> ClassifyInputs {
         ClassifyInputs {
             n_join_tables: n,
-            has_group_by: group,
-            has_aggregate: agg,
+            has_group_by: gb,
             all_aggregates_splittable: splittable,
+            has_aggregate: agg,
         }
     }
 
     #[test]
-    fn classify_scalar_agg_on_binary_join() {
-        // COUNT(*) FROM f JOIN p WHERE ...
+    fn binary_scalar_agg() {
         assert_eq!(
-            classify(&inputs(2, false, true, true)),
+            classify(&inputs(2, true, false, true)),
             MppPlanShape::ScalarAggOnBinaryJoin
         );
     }
 
     #[test]
-    fn classify_groupby_agg_on_binary_join() {
-        // SELECT k, COUNT(*) FROM f JOIN p GROUP BY k
+    fn binary_groupby_agg() {
         assert_eq!(
             classify(&inputs(2, true, true, true)),
             MppPlanShape::GroupByAggOnBinaryJoin
@@ -545,8 +135,7 @@ mod tests {
     }
 
     #[test]
-    fn classify_groupby_agg_single_table() {
-        // SELECT k, COUNT(*) FROM t GROUP BY k
+    fn single_groupby() {
         assert_eq!(
             classify(&inputs(1, true, true, true)),
             MppPlanShape::GroupByAggSingleTable
@@ -554,65 +143,26 @@ mod tests {
     }
 
     #[test]
-    fn classify_join_only() {
-        // SELECT * FROM f JOIN p — no aggregate at all.
+    fn join_only() {
         assert_eq!(
             classify(&inputs(2, false, false, true)),
             MppPlanShape::JoinOnly
         );
-        // A CLASSIFY_OUTPUT_BY with no aggregate is still join-only; GROUP BY
-        // without aggregates is unusual SQL but should not elevate to group-by
-        // aggregate shape because there's no partial state to shuffle.
-        assert_eq!(
-            classify(&inputs(2, true, false, true)),
-            MppPlanShape::JoinOnly
-        );
     }
 
     #[test]
-    fn classify_rejects_unsplittable_aggregates() {
-        // ARRAY_AGG, COUNT(DISTINCT), etc. aren't safe to Partial/Final split.
+    fn non_splittable_aggregate_is_ineligible() {
         assert_eq!(
-            classify(&inputs(2, false, true, false)),
-            MppPlanShape::Ineligible
-        );
-        assert_eq!(
-            classify(&inputs(2, true, true, false)),
+            classify(&inputs(2, true, false, false)),
             MppPlanShape::Ineligible
         );
     }
 
     #[test]
-    fn classify_rejects_single_table_scalar() {
-        // Single-table scalar aggregates don't benefit from MPP — the
-        // aggregate is already O(rows/workers) via PG's native parallel
-        // scan. No shuffle needed.
+    fn three_table_join_is_ineligible() {
         assert_eq!(
-            classify(&inputs(1, false, true, true)),
+            classify(&inputs(3, true, false, true)),
             MppPlanShape::Ineligible
         );
     }
-
-    #[test]
-    fn classify_rejects_no_tables() {
-        assert_eq!(
-            classify(&inputs(0, false, false, true)),
-            MppPlanShape::Ineligible
-        );
-    }
-
-    #[test]
-    fn classify_rejects_multi_table_join() {
-        // 3+ table joins aren't wired yet (scope = milestone 1 binary join).
-        // Extend `classify` when milestone-2 adds them.
-        assert_eq!(
-            classify(&inputs(3, false, true, true)),
-            MppPlanShape::Ineligible
-        );
-    }
-
-    // `shuffles_required` was retired in P4 in favor of
-    // `walker::cut_count_for_shape`; the equivalent coverage lives in
-    // `walker::tests::cut_count_matches_expected_cuts_len` plus
-    // `expected_cuts_stamp_correct_stage_ids`.
 }
